src/backend/requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.96.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
orjson==3.9.0
pandas==2.0.2
pydantic==1.10.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rdflib==6.3.2
requests==2.31.0
scikit-learn==1.2.2
scikit-surprise==1.1.3
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SPARQLWrapper==2.0.0
starlette==0.27.0
threadpoolctl==3.1.0
tmdbsimple==2.9.1
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
